Added CodeGraphContext to readme
CodeGraphContext is an MCP server that indexes local code into a graph database to provide context to AI assistants. It has over 7k downloads and 200+ stars within a period of just 2 months. It is already bringing craze to prominent reddit and linkedin communities with a post on r/MCP occupying the "3rd best post" ever.